trivially small upgrade
(This upgrade is hardly worth the bytes it is written on :-).)

I think 'applying' a function is more idiomatic than 'calling' a function when talking about FP. It's not the hill I want to die on nor the yak I want to shave. I can't believe I have invested even the time to write this comment about something so pedantic that even the Angels dancing on their pinheads are shaking their heads and tutting at me.