Fix a typo
Additionaly, I do not see the point of showing us the definition of normalize2: except for the docstring it's the same as the function normalize define just before. Did you mean to show the contrast with a function that takes two arguments x and y as opposed to a function that takes one argument (a vector)?